fingerprinter/requirements.txt
tensorflow-gpu
opencv-python
efficientnet
